+++ models/Context.php
@@ -121,6 +121,11 @@ class Context extends Component
                 }
             }
         }
+        foreach ($this->interfaces as $interface) {
+            foreach ($interface->parentInterfaces as $pInterface) {
+                $this->interfaces[$pInterface]->implementedBy[] = $interface->name;
+            }
+        }
         // inherit docs
         foreach ($this->classes as $class) {
             $this->inheritDocs($class);
@@ -128,6 +133,9 @@ class Context extends Component
         // inherit properties, methods, contants and events to subclasses
         foreach ($this->classes as $class) {
             $this->updateSubclassInheritance($class);
+        }
+        foreach ($this->interfaces as $interface) {
+            $this->updateSubInterfaceInheritance($interface);
         }
         // add properties from getters and setters
         foreach ($this->classes as $class) {
@@ -164,6 +172,21 @@ class Context extends Component
             $subclass->properties = array_merge($class->properties, $subclass->properties);
             $subclass->methods = array_merge($class->methods, $subclass->methods);
             $this->updateSubclassInheritance($subclass);
+        }
+    }
+
+    /**
+     * Add methods to subinterfaces
+     * @param InterfaceDoc $class
+     */
+    protected function updateSubInterfaceInheritance($interface)
+    {
+        foreach ($interface->implementedBy as $subInterface) {
+            if (isset($this->interfaces[$subInterface])) {
+                $subInterface = $this->interfaces[$subInterface];
+                $subInterface->methods = array_merge($interface->methods, $subInterface->methods);
+                $this->updateSubInterfaceInheritance($subInterface);
+            }
         }
     }