+++ templates/html/views/methodDetails.php
@@ -37,6 +37,8 @@ ArrayHelper::multisort($methods, 'name');
         <?= $method->name ?>()
         <span class="detail-header-tag small">
             <?= $method->visibility ?>
+            <?= $method->isAbstract ? 'abstract' : '' ?>
+            <?= $method->isStatic ? 'static' : '' ?>
             method
             <?php if (!empty($method->since)): ?>
                 (available since version <?= $method->since ?>)
@@ -53,7 +55,7 @@ ArrayHelper::multisort($methods, 'name');
     </div>
 
     <table class="detail-table table table-striped table-bordered table-hover">
-        <tr><td colspan="3" class="signature"><?= $renderer->renderMethodSignature($method) ?></td></tr>
+        <tr><td colspan="3" class="signature"><?= $renderer->renderMethodSignature($method, $type) ?></td></tr>
         <?php if (!empty($method->params) || !empty($method->return) || !empty($method->exceptions)): ?>
             <?php foreach ($method->params as $param): ?>
                 <tr>
@@ -65,7 +67,7 @@ ArrayHelper::multisort($methods, 'name');
             <?php if (!empty($method->return)): ?>
                 <tr>
                   <th class="param-name-col">return</th>
-                  <td class="param-type-col"><?= $renderer->createTypeLink($method->returnTypes) ?></td>
+                  <td class="param-type-col"><?= $renderer->createTypeLink($method->returnTypes, $type) ?></td>
                   <td class="param-desc-col"><?= ApiMarkdown::process($method->return, $type) ?></td>
                 </tr>
             <?php endif; ?>