+++ CHANGELOG.md
@@ -1,10 +1,15 @@
 Yii Framework 2 apidoc extension Change Log
 ===========================================
 
-2.0.4 under development
+2.0.5 under development
 -----------------------
 
-- no changes in this release.
+
+2.0.4 May 10, 2015
+------------------
+
+- Bug #3: Interface documentation did not show inheritance (cebe)
+- Enh: Added ability to set pageTitle from command line (unclead)
 
 
 2.0.3 March 01, 2015